backendv2/requirements.txt
psycopg2-binary
python-dotenv
llama-index
llama-index-retrievers-bm25
google-adk